Move retrain button into id. Extract Reports into component
